Added local deployment startup to worker connection tests
